gui in progress and changed few things

+++ trunk/src/main/java/cz/muni/fi/pb138/SceneParser.java
@@ -4,11 +4,10 @@
  */
 package cz.muni.fi.pb138;
 
-import java.io.File;
-import java.io.IOException;
-import java.net.URI;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+import org.w3c.dom.Document;
+import org.w3c.dom.Node;
+import org.xml.sax.SAXException;
+
 import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
@@ -16,9 +15,11 @@ import javax.xml.xpath.XPath;
 import javax.xml.xpath.XPathConstants;
 import javax.xml.xpath.XPathExpressionException;
 import javax.xml.xpath.XPathFactory;
-import org.w3c.dom.Document;
-import org.w3c.dom.Node;
-import org.xml.sax.SAXException;
+import java.io.File;
+import java.io.IOException;
+import java.net.URI;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  *
@@ -45,7 +46,7 @@ public class SceneParser {
         doc = builder.parse(uri.toString());
     }
     
-    public GameScene parseScene(int id) {
+    public GameScene parseScene(long id) {
         String xPathScene; 
         Node node;
         try {
